% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicanalysisBundle_class.R
\docType{data}
\name{Annotations}
\alias{Annotations}
\title{Annotations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Annotations
}
\description{
Annotations
}
\section{Fields}{

\describe{
\item{\code{foo}}{some explanation}

\item{\code{foo2}}{anotther explanation}
}}
\section{Methods}{

\describe{
\item{\code{anyMethod(x)}}{do something}
}
}
\examples{
\dontrun{
}
}
\keyword{datasets}

