% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_method.R
\docType{methods}
\name{cooccurrences,TopicModel-method}
\alias{cooccurrences,TopicModel-method}
\title{Calculate cooccurrences for a topicmodel}
\usage{
\S4method{cooccurrences}{TopicModel}(.Object, k, regex = NULL, docs = NULL,
  progress = TRUE, verbose = TRUE)
}
\arguments{
\item{.Object}{the object}

\item{k}{the k first topics to consider}

\item{verbose}{logical}
}
\description{
Calculate cooccurrences for a topicmodel
}
\examples{
\dontrun{
setwd("/Users/blaette/Lab/repos/arenen/analysis/corpusConsolidation/BT")
lda <- readRDS("lda_bt_arbeit_50.RData")
topic_cooccurrencs <- cooccurrences(lda, k=3)
dt <- subset(topic_cooccurrences, chisquare >= threshold, select=c("x", "y"))
igraph::graph_from_data_frame(dt)
}
}

