% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BE_exclude}
\alias{BE_exclude}
\alias{BE_labels}
\alias{BE_lda}
\alias{SL_exclude}
\alias{SL_labels}
\alias{SL_lda}
\title{Sample LDA topicmodels}
\format{Topicmodels for German regional parliaments (n = 250).}
\usage{
BE_exclude

BE_labels

BE_lda

SL_exclude

SL_labels

SL_lda
}
\description{
Sample LDA topicmodels
}
\keyword{datasets}
