% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldaBundle.R
\docType{class}
\name{ldaBundle-class}
\alias{ldaBundle-class}
\alias{ldaBundle}
\alias{mean,ldaBundle-method}
\alias{plot,ldaBundle,ANY-method}
\alias{max,ldaBundle-method}
\title{calculate and compare a set of topicmodels}
\usage{
ldaBundle(dtm, ks, method = "Gibbs", burnin = 1000, iter = 1000,
  keep = 50, mc = 2, verbose = TRUE)

\S4method{mean}{ldaBundle}(x)

\S4method{plot}{ldaBundle,ANY}(x, ks, burnin, keep)

\S4method{max}{ldaBundle}(x)
}
\arguments{
\item{dtm}{dtm}

\item{ks}{numeric vector, number of topics}

\item{method}{method (defaults to "Gibbs")}

\item{burnin}{foo}

\item{iter}{number of iterations}

\item{keep}{keep}

\item{mc}{number of cores}

\item{verbose}{verbose}

\item{x}{x}
}
\description{
calculate and compare a set of topicmodels
}
