% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_method.R
\docType{methods}
\name{read,TopicModel-method}
\alias{read,TopicModel-method}
\title{read-method for TopicModel}
\usage{
\S4method{read}{TopicModel}(.Object, partitionObject, noTopics = 3,
  noToken = 20)
}
\arguments{
\item{.Object}{object of the TopicModel class, or inheriting from it}

\item{partitionObject}{partition to be read}

\item{noTopics}{the number of the most prevalant topics in a text that will be highlighted}

\item{noToken}{the number of tokens to be be highlighted}
}
\description{
foo
}

