% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentenceSimilarity.R
\docType{data}
\name{SentenceSimilarity}
\alias{SentenceSimilarity}
\title{Get similar sentences.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SentenceSimilarity
}
\description{
Get similar sentences.
}
\section{Fields}{

\describe{
\item{\code{partition}}{a partition}
}}

\section{Methods}{

\describe{
  \item{\code{getSentenceVectors(method = c("pos", "openNLP"), f = mean, stopwords = tm::stopwords("de"), tfidf = FALSE, verbose = TRUE)}}{}
  \item{\code{}}{}
}
}

\examples{
\dontrun{
library(R6)
library(NLP)
library(openNLP)
library(openNLPmodels.de)
library(polmineR)

filename <- "/Users/blaette/Lab/tmp/word2vec/word2vec/data/plprbt_word2vec.csv"
W2V <- Word2Vec$new()
W2V$load(filename)
merkel <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_year = "2015")
S <- SentenceSimilarity$new()
S$partition <- merkel
S$word2vec <- W2V$vector_matrix
S$threads <- 3
S$getSentenceVectors(method = "pos", f = sum)
S$getSimilarityMatrix()

i <- 549
S$sentences[[i]]
sentence <- c("Kriminelle", "Ausländer", "raus", "und", "zwar", "schnell")
S$getSentenceSimilarity(sentence)
}
}
\keyword{datasets}
