% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny}
\alias{shiny}
\title{start shiny application for topic analysis}
\usage{
shiny(dev = TRUE)
}
\arguments{
\item{dev}{logical, if TRUE, start shiny app in dev version of the package}
}
\description{
start shiny application for topic analysis
}
