% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicanalysisBundle_class.R
\docType{data}
\name{TopicanalysisBundle}
\alias{TopicanalysisBundle}
\title{TopicanalysisBundle}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TopicanalysisBundle
}
\description{
TopicanalysisBundle
}
\section{Fields}{

\describe{
\item{\code{foo}}{something}
}}
\examples{
\dontrun{
B <- TopicanalysisBundle$new(topicmodels = TAs)
B$makeMatrix(big = T)
saveRDS(B$termTopicMatrix, file="/Users/blaette/Lab/repos/diffusion/matrix.RData")
B$termTopicMatrix <- readRDS("/Users/blaette/Lab/repos/diffusion/matrix.RData")
B$similarity <- readRDS("/Users/blaette/Lab/repos/diffusion/simil.RData")
B$doClustering()
B$getClusters(h=1)
}
}
\keyword{datasets}

