% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicmodelsCmd_function.R
\name{topicmodelsCmd}
\alias{topicmodelsCmd}
\title{generate command to run lds topicmodelling from command line}
\usage{
topicmodelsCmd(method = "lda", infile, outfile, k)
}
\arguments{
\item{method}{either ctm or lda}

\item{infile}{input file}

\item{outfile}{output file (full path)}

\item{k}{number of topics}
}
\description{
generate command to run lds topicmodelling from command line
}

