% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Word2Vec.R
\docType{data}
\name{Word2Vec}
\alias{Word2Vec}
\title{Use word2vec.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Word2Vec
}
\description{
Use word2vec.
}
\examples{
\dontrun{
filename <- "/Users/blaette/Lab/tmp/word2vec/word2vec/data/plprbt_word2vec.csv"
W2V <- Word2Vec$new()
}
}
\keyword{datasets}

