% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.zoo_method.R
\docType{methods}
\name{as.zoo}
\alias{as.zoo}
\alias{as.zoo,LDA-method}
\title{aggregate topic occurrence according to time}
\usage{
as.zoo(x, ...)

\S4method{as.zoo}{LDA}(x, k = 3,
  regex = "^.*?(\\\\d{4}-\\\\d{2}-\\\\d{2}).*?$", select = NULL,
  aggregation = NULL)
}
\arguments{
\item{x}{a LDA object}

\item{...}{further arguments}

\item{k}{include topics into aggregation that are among the top k topics}

\item{regex}{a regex to extract the date from the document name}

\item{select}{topics to select}

\item{aggregation}{either "year", "quarter" or month}
}
\description{
aggregate topic occurrence according to time
}

