% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.zoo.R
\docType{methods}
\name{as.zoo}
\alias{as.zoo}
\alias{as.zoo,LDA-method}
\title{Generate Time Series Object From Topicmodel.}
\usage{
as.zoo(x, ...)

\S4method{as.zoo}{LDA}(x, k = 3,
  regex = "^.*?(\\\\d{4}-\\\\d{2}-\\\\d{2}).*?$", select = NULL,
  aggregation = NULL)
}
\arguments{
\item{x}{A topicmodel inheriting from class \code{LDA}.}

\item{...}{further arguments}

\item{k}{Number of top topics considered during aggregation.}

\item{regex}{Regular expression to extract the date from the document name.}

\item{select}{topics to select}

\item{aggregation}{either "year", "quarter" or month, if NULL, no aggregation is performed.}
}
\description{
Generate Time Series Object From Topicmodel.
}
