% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mallet.R
\docType{methods}
\name{as.mallet}
\alias{as.mallet}
\alias{as.mallet,partition_bundle-method}
\alias{malletCmd}
\alias{malletImport}
\alias{store,jobjRef-method}
\title{Interface to mallet topicmodelling.}
\usage{
as.mallet(.Object, ...)

\S4method{as.mallet}{partition_bundle}(.Object, pAttribute = "word",
  termsToDrop = tm::stopwords("de"), mc = TRUE, verbose = TRUE)

malletCmd(sourcefile, targetDir = "/Users/blaette/Lab/tmp/mallet_result",
  topwords = 50, topics = 50, iterations = 2000, threads = 1)

malletImport(dir = "/Users/blaette/Lab/tmp/mallet_result")

\S4method{store}{jobjRef}(.Object, filename = tempfile())
}
\arguments{
\item{.Object}{\code{partition}, \code{partition_bundle}, or \code{rJava} object}

\item{...}{further parameters}

\item{pAttribute}{the pAttribute to use, typically "word" or "lemma"}

\item{termsToDrop}{stopwords}

\item{mc}{whether to use multicore}

\item{verbose}{logical, whether to be verbose}

\item{sourcefile}{...}

\item{targetDir}{...}

\item{topwords}{...}

\item{topics}{...}

\item{iterations}{...}

\item{threads}{...}

\item{dir}{...}

\item{filename}{where to store the Java-object}
}
\description{
Turn partitionBundle into a mallet object (instance list). The typical 
workflow will be as follows (see example):
(1) Turn partitionBundle-object into mallet instance list, (2a) store the
resulting jobjRef-object locally using store-method, create mallet command
with malletCmd, and run mallet from the command line; alternativelly (2b) run
mallet topic modelling from within R.
}
\examples{
\dontrun{
use("polmineR.sampleCorpus")
bt2009 <- partition("PLPRBTTXT", text_year="2009")
bt2009bundle <- partitionBundle(bt2009, sAttribute= "text_protocol_no", pAttribute = NULL)
instanceList <- as.mallet(bt2009bundle)
topic.model <- MalletLDA(num.topics = 20)
topic.model$loadDocuments(instanceList)
topic.model$setAlphaOptimization(20, 50)
topic.model$train(2000)

fname <- store(instanceList, filename = NULL) # output to a tempfile
system(malletCmd(fname))
}
}
