% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{methods}
\name{get_highlight_list}
\alias{get_highlight_list}
\alias{read,TopicModel-method}
\title{Fulltext Inspection for TopicModel.}
\usage{
get_highlight_list(.Object, partition_obj, no_topics = 3L,
  no_token = 20L)

\S4method{read}{TopicModel}(.Object, partition_obj, no_topics = 3L,
  no_token = 20L)
}
\arguments{
\item{.Object}{Object of class \code{TopicModel}, or a class inheriting from
it.}

\item{partition_obj}{A \code{partition} object, the partition to display.}

\item{no_topics}{The number of the most prevalant topics in a text that will be highlighted.}

\item{no_token}{An \code{integer} value, the number of tokens indicative of a
topic according to the input topic model to be be highlighted}
}
\description{
Highlight the vocabulary indicative of the top terms (number specified by
\code{no_topics}) for the top topics (number specified by
\code{no_topics}present in a \code{partition}.
}
\examples{
data(BE_lda)
data(BE_labels)
data(BE_exclude)

BE <- Topicanalysis$new(topicmodel = BE_lda)
BE$labels <- BE_labels
BE$type <- "plpr_partition"

topic_flucht <- 125L
topic_integration <- 241
BE$docs(x = 125L, y = 241L)

# As sample data, we use a corpus of plenary debates in the 
# 'Berliner Abgeordnetenhaus'. It is somewhat big and not included
# in the package by default. 

\dontrun{
polmineR::use("topicanalysis")
if (!"BE" \%in\% corpus()$corpus){
  cwbtools::corpus_install(
    pkg = "topicanalysis",
    tarball = "http://polmine.sowi.uni-due.de/corpora/cwb/berlin/be.tar.gz"
  )
  polmineR::use("topicanalysis") # to activate corpus 'BE'
}

p <- partition("BE", date = "2005-04-28", who = "Körting")
p <- polmineR::as.speeches(p, s_attribute_name = "who")[[4]]

read(BE_lda, p, no_token = 150)
}
}
