% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topicanalysis.R
\name{Topicanalysis}
\alias{Topicanalysis}
\title{Topicanalysis Class.}
\description{
Analyse topicmodels.
}
\section{Arguments}{

\describe{
  \item{new}{New value for a label or a category.}
  \item{n}{Number of a topic.}
  \item{n_words}{An \code{integer}, the number of words to be displayed in a wordcloud.}
  \item{x}{Number or name of a topics.}
  \item{y}{Number or name of a topic cooccurring with x.}
  \item{k}{Number of top topics of a document to consider.}
  \item{exclude}{A \code{logical} value, whether to to exclude topics
  earmarked in logical vector in field exclude.}
  \item{aggregation}{Level of aggregation of \code{as.zoo} method.}
  \item{...}{Further parameters passed to worker function (such as
  \code{wordcloud::wordcloud} when calling \code{$wordcloud()}, for
  instance).}
  \item{regex}{A regular expression that will limit the evaluation to those documents only
  that are matched by the regular expression.}
}
}

\section{Methods}{

\describe{
  \item{\code{$initialize(topicmodel)}}{Instantiate new \code{Topicanalysis}
  object. Upon initialization, labels will be the plain numbers of the
  topics, all exclude values are \code{FALSE}.}
  \item{\code{$cooccurrences(k = 3, regex = NULL, docs = NULL, renumber = NULL,
  progress = TRUE, exclude = TRUE)}}{Get cooccurrences of topics. Arguments are
  documented with the S4 cooccurrences-method for TopicModel-objects.}
  \item{\code{$relabel(n, new)}}{Relabel topic \code{n}, assigning new label
  \code{new}.}
  \item{\code{$add_category(new)}}{Add \code{new}, a \code{character} vector
  as a new category to the \code{character} vector in the field
  \code{category}.}
  \item{\code{$ignorance(n, new)}}{Exclude topic \code{n} (i.e. add to ignore).}
  \item{\code{$wordcloud(n, n = 50, ...)}}{Generate wordcloud for topic
  \code{n}, with \code{n_words} words. Further arguments can be passed into
  \code{wordcloud::wordcloud} usint the three dots.}
  \item{\code{$docs(x, y = NULL, n = 3L, s_attributes = NULL)}}{Get documents
  where topic \code{x} occurrs among the top \code{n} topics. If \code{y} is 
  provided, documents are returned where \code{x} and \code{y} are among the
  \code{n} top topics. If \code{x} or \code{y} are provided as a \code{character}
  vector, the method will look up this label in the \code{labels} field.}
  \item{\code{$read(x, n = 3, no_token = 100)}}{Read document \code{x},
  highlighting the number of topics specified by \code{n}, indicated by
  \code{no_token}.}
  \item{\code{$as.zoo(x = NULL, y = NULL, k = 3, exclude = TRUE, aggregation
  = c(NULL, "month", "quarter", "year"))}}{Generate \code{zoo} object from
  topicmodel.}
  \item{\code{$compare(x, ...)}}{Compare the similarity of two topicmodels.}
  \item{\code{$find_topics(x, n = 100, word2vec = NULL)}}{Find a topic.}
}
}

\examples{
data(BE_lda)
data(BE_labels)
data(BE_exclude)

BE <- Topicanalysis$new(topicmodel = BE_lda)
BE$labels <- BE_labels
BE$exclude <- BE_exclude
BE$exclude <- grepl("^\\\\((split|)\\\\)$", BE$labels)
BE$name <- "Berlin"
BE$type <- "plpr_partition"

z <- BE$as.zoo(x = "Flucht, Asyl, vorläufiger Schutz", aggregation = "year")
plot(z)

y <- BE$as.zoo(
  x = grep("Asyl", BE_labels),
  y = grep("Europ", BE_labels),
  aggregation = "year"
)
plot(y)

BE$exclude <- grepl("^\\\\(.*?\\\\)$", BE$labels)
dt <- BE$cooccurrences(k = 3L, exclude = TRUE)
dt_min <- dt[chisquare >= 10.83]


if (requireNamespace("igraph")){
g <- igraph::graph_from_data_frame(
  d = data.frame(
    from = dt_min[["a_label"]],
    to = dt_min[["b_label"]],
    n = dt_min[["count_coi"]],
    stringsAsFactors = FALSE
  ),
  directed = TRUE
)
g <- igraph::as.undirected(g, mode = "collapse")
if (interactive()){
  igraph::plot.igraph(
    g, shape = "square", vertex.color = "steelblue",
    label = igraph::V(g)$name, label.family = 11, label.cex = 0.5
  )
}
}

topic_flucht <- 125L
topic_integration <- 241
BE$docs(x = "Flucht, Asyl, vorläufiger Schutz")
BE$docs(x = grep("Flucht", BE$labels))
BE$docs(x = 125L)
docs <- BE$docs(x = 125L, y = 241L)

\dontrun{
li <- lapply(
  docs, 
  function(doc){
    polmineR::as.speeches(
      polmineR::partition(
        "BE",
        who = gsub("^(.*?)_.*$", "\\\\1", doc),
        date = gsub("^.*(\\\\d{4}-\\\\d{2}-\\\\d{2})_\\\\d+$", "\\\\1", doc)
      ),
      s_attribute_name = "who"
    )[[as.integer(gsub("^.*?_(\\\\d+)$", "\\\\1", doc))]]
})
BE$bundle <- as.partition_bundle(li)

read(BE$topicmodel, BE$bundle[[1]], n = 3L, no_token = 250)
read(BE$topicmodel, BE$bundle[[2]], n = 3L, no_token = 250)
read(BE$topicmodel, BE$bundle[[3]], n = 3L, no_token = 250)
for (doc in docs){
  print(doc)
  p <- BE$bundle[[doc]]
  read(BE$topicmodel, p, n = 3L, no_token = 250)
  readline(prompt = "Hit any key to continue.")
}
}

#############################

data(SL_lda)
data(SL_labels)
data(SL_exclude)

SL <- Topicanalysis$new(topicmodel = SL_lda)
SL$labels <- SL_labels
SL$exclude <- SL_exclude
SL$exclude <- grepl("^\\\\((split|)\\\\)$", SL$labels)
SL$name <- "Hamburg"

cp_1 <- BE$compare(SL, BE)
cp_2 <- BE$compare(SL, BE)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{topicmodel}}{A topicmodel of class \code{TopicModel}, generated from
package \code{topicmodels}.}

\item{\code{posterior}}{Slot to store posterior, not used at this point.}

\item{\code{terms}}{The \code{matrix} with the terms of a topicmodel. Keeping the
terms may speed up subsequent operations.}

\item{\code{topics}}{The \code{matrix} with the topics present in documents. Keeping
this matrix may speed up subsequent operations.}

\item{\code{bundle}}{A \code{partition_bundle}, required to use method \code{read}
to access full text.}

\item{\code{labels}}{A \code{character} vector, labels for the topics.}

\item{\code{name}}{A name for the \code{Topicanalysis} object. Useful if combining
several objects into a bundle.}

\item{\code{categories}}{A \code{character} vector with categories.}

\item{\code{grouping}}{Not used at this stage.}

\item{\code{exclude}}{Topics to exclude from further analysis.}

\item{\code{type}}{Corpus type, necessary for applying correct template for fulltext output.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Topicanalysis$new()}}
\item \href{#method-relabel}{\code{Topicanalysis$relabel()}}
\item \href{#method-add_category}{\code{Topicanalysis$add_category()}}
\item \href{#method-ignorance}{\code{Topicanalysis$ignorance()}}
\item \href{#method-cooccurrences}{\code{Topicanalysis$cooccurrences()}}
\item \href{#method-wordcloud}{\code{Topicanalysis$wordcloud()}}
\item \href{#method-docs}{\code{Topicanalysis$docs()}}
\item \href{#method-read}{\code{Topicanalysis$read()}}
\item \href{#method-as.zoo}{\code{Topicanalysis$as.zoo()}}
\item \href{#method-compare}{\code{Topicanalysis$compare()}}
\item \href{#method-find_topics}{\code{Topicanalysis$find_topics()}}
\item \href{#method-clone}{\code{Topicanalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$new(topicmodel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-relabel"></a>}}
\subsection{Method \code{relabel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$relabel(n, new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_category"></a>}}
\subsection{Method \code{add_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$add_category(new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ignorance"></a>}}
\subsection{Method \code{ignorance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$ignorance(n, new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cooccurrences"></a>}}
\subsection{Method \code{cooccurrences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$cooccurrences(
  k = 3,
  regex = NULL,
  docs = NULL,
  renumber = NULL,
  progress = TRUE,
  verbose = FALSE,
  exclude = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wordcloud"></a>}}
\subsection{Method \code{wordcloud()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$wordcloud(n, n_words = 50, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docs"></a>}}
\subsection{Method \code{docs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$docs(x, y = NULL, n = 3L, s_attributes = NULL, regex = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$read(x, n = 3L, no_token = 100L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.zoo"></a>}}
\subsection{Method \code{as.zoo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$as.zoo(
  x = NULL,
  y = NULL,
  k = 3L,
  exclude = TRUE,
  aggregation = c(NULL, "month", "quarter", "year")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compare"></a>}}
\subsection{Method \code{compare()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$compare(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_topics"></a>}}
\subsection{Method \code{find_topics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$find_topics(x, n = 100, word2vec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topicanalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
