% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topicanalysis.R
\docType{data}
\name{Topicanalysis}
\alias{Topicanalysis}
\title{Topicanalysis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Topicanalysis
}
\arguments{
\item{new}{New value for a label or a category.}

\item{n}{Number of a topic.}

\item{n_words}{An integer, the number of words to be displayed in a wordcloud.}

\item{x}{Number or name of a topics.}

\item{y}{Number or name of a topic cooccurring with x.}

\item{k}{Number of top topics in a document considered.}

\item{exclude}{A logical value, whether to to exclude topics earmarked in
logical vector in field exclude.}

\item{aggregation}{Level of aggregation of \code{as.zoo} method.}

\item{...}{Further parameters passed to worker function (\code{wordcloud}, for instance).}
}
\description{
Analyse topicmodels.
}
\section{Fields}{

\describe{
\item{\code{topicmodel}}{A topicmodel of class \code{TopicModel}, generated from package \code{topicmodels}.}

\item{\code{posterior}}{Slot to store posterior, not used at this point.}

\item{\code{terms}}{The \code{matrix} with the terms of a topicmodel; stored to .}

\item{\code{bundle}}{A \code{partitionBundle}, required to use method \code{read} to access full text.}

\item{\code{labels}}{A character vector, labels for the topics.}

\item{\code{name}}{A name for the \code{Topicanalysis} object. Useful if combining
several objects into a bundle.}

\item{\code{categories}}{A character vector with categories.}

\item{\code{grouping}}{Not used at this stage.}

\item{\code{exclude}}{Topics to exclude from further analysis.}

\item{\code{type}}{Corpus type, necessary for applying correct template for fulltext output.}
}}

\section{Methods}{

\describe{
  \item{\code{$initialize(topicmodel)}}{Instantiate new \code{Topicanalysis}
  object. Upon initialization, labels will be the plain numbers of the
  topics, all exclude values are \code{FALSE}.}
  \item{\code{$cooccurrences(k = 3, regex = NULL, docs = NULL, renumber = NULL,
  progress = TRUE, exclude = TRUE)}}{ Get cooccurrences of topics. Params are
  documented with the S4 cooccurrences-method for TopicModel-objects. }
  \item{\code{$relabel(n, new)}}{Relabel topic \code{n}, assigning new label \code{new}.}
  \item{\code{$add_category(new)}}{Add a new category.}
  \item{\code{$ignorance(n, new)}}{Exclude topic \code{n} (i.e. add to ignore).}
  \item{\code{$cooccurrence(k = 3, regex = NULL, docs = NULL, renumber = NULL, progress = TRUE, exclude = TRUE)}}{Get cooccurrences of topics.}
  \item{\code{$wordcloud(x = 1, n = 50, ...)}}{Generate wordcloud for topic \code{x}, with \code{n} words.}
  \item{\code{$docs(x, y = NULL, n = 3, sAttributes = NULL)}}{Get documents where topic \code{x} occurrs among the top \code{n} topics.}
  \item{\code{$read(x, n = 3, noToken = 100)}}{Read stuff.}
  \item{\code{$as.zoo(x = NULL, y = NULL, k = 3, exclude = TRUE, aggregation
  = c(NULL, "month", "quarter", "year"))}}{Generate \code{zoo} object from topicmodel.}
  \item{\code{$compare(x, ...)}}{xxx}
  \item{\code{$find_topics(x, n = 100, word2vec = NULL)}}{Find a topic.}
}
}

\keyword{datasets}
