% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topicanalysis_class.R
\docType{data}
\name{Topicanalysis}
\alias{Topicanalysis}
\title{Topicanalysis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Topicanalysis
}
\description{
Analyse topicmodels.
}
\examples{
\dontrun{
library(polmineR.topics)
setwd("/Users/blaette/Lab/repos/arenen/analysis/policy_events_paper")
options(polmineR$meta=c("text_name", "text_party", "text_date")))
speeches <- readRDS("bundle/speeches.RData")
lda <- readRDS("lda/lda_bt_migration.RData")
TA <- Topicanalysis$new(topicmodel=lda, bundle=speeches)
TA$labels # see the result
saveRDS(TA$labels, file="lda_bt_migration_50_labels.RData")
TA$labels <- readRDS("lda_bt_migration_50_labels.RData")

TA$as.zoo(period="year")
TA$cooccurring()
}
}
\keyword{datasets}

