% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicanalysisBundle.R
\docType{data}
\name{TopicanalysisBundle}
\alias{TopicanalysisBundle}
\title{TopicanalysisBundle}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TopicanalysisBundle
}
\description{
TopicanalysisBundle
}
\section{Fields}{

\describe{
\item{\code{foo}}{something}
}}

\keyword{datasets}
