% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicanalysisBundle.R
\name{TopicanalysisBundle}
\alias{TopicanalysisBundle}
\title{TopicanalysisBundle}
\description{
TopicanalysisBundle

TopicanalysisBundle
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{foo}}{something}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{foo}}{something}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-length}{\code{TopicanalysisBundle$length()}}
\item \href{#method-names}{\code{TopicanalysisBundle$names()}}
\item \href{#method-restore}{\code{TopicanalysisBundle$restore()}}
\item \href{#method-store}{\code{TopicanalysisBundle$store()}}
\item \href{#method-getTerms}{\code{TopicanalysisBundle$getTerms()}}
\item \href{#method-makeTermTopicMatrix}{\code{TopicanalysisBundle$makeTermTopicMatrix()}}
\item \href{#method-getTopicSimilarity}{\code{TopicanalysisBundle$getTopicSimilarity()}}
\item \href{#method-getTopicDistance}{\code{TopicanalysisBundle$getTopicDistance()}}
\item \href{#method-doClustering}{\code{TopicanalysisBundle$doClustering()}}
\item \href{#method-plot}{\code{TopicanalysisBundle$plot()}}
\item \href{#method-getTopicClusters}{\code{TopicanalysisBundle$getTopicClusters()}}
\item \href{#method-makeOverviewTopicClusters}{\code{TopicanalysisBundle$makeOverviewTopicClusters()}}
\item \href{#method-plotClusterDendrogram}{\code{TopicanalysisBundle$plotClusterDendrogram()}}
\item \href{#method-plotHeatmap}{\code{TopicanalysisBundle$plotHeatmap()}}
\item \href{#method-labelTopicClusters}{\code{TopicanalysisBundle$labelTopicClusters()}}
\item \href{#method-assignLabels}{\code{TopicanalysisBundle$assignLabels()}}
\item \href{#method-query}{\code{TopicanalysisBundle$query()}}
\item \href{#method-as.zoo}{\code{TopicanalysisBundle$as.zoo()}}
\item \href{#method-clone}{\code{TopicanalysisBundle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-names"></a>}}
\subsection{Method \code{names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restore"></a>}}
\subsection{Method \code{restore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$restore(x = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\subsection{Method \code{store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$store(verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTerms"></a>}}
\subsection{Method \code{getTerms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$getTerms(k = 100, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeTermTopicMatrix"></a>}}
\subsection{Method \code{makeTermTopicMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$makeTermTopicMatrix(
  output = "matrix",
  minimize = FALSE,
  verbose = TRUE,
  progress = TRUE,
  mc = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTopicSimilarity"></a>}}
\subsection{Method \code{getTopicSimilarity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$getTopicSimilarity(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTopicDistance"></a>}}
\subsection{Method \code{getTopicDistance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$getTopicDistance(
  chunks = 1,
  method = "binary",
  mc = FALSE,
  progress = TRUE,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doClustering"></a>}}
\subsection{Method \code{doClustering()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$doClustering()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$plot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTopicClusters"></a>}}
\subsection{Method \code{getTopicClusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$getTopicClusters(h, k = 25)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeOverviewTopicClusters"></a>}}
\subsection{Method \code{makeOverviewTopicClusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$makeOverviewTopicClusters(n = 10, terms = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotClusterDendrogram"></a>}}
\subsection{Method \code{plotClusterDendrogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$plotClusterDendrogram(topicNo, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHeatmap"></a>}}
\subsection{Method \code{plotHeatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$plotHeatmap(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-labelTopicClusters"></a>}}
\subsection{Method \code{labelTopicClusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$labelTopicClusters(min = 0, max = 25, heatmap = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignLabels"></a>}}
\subsection{Method \code{assignLabels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$assignLabels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$query(query)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.zoo"></a>}}
\subsection{Method \code{as.zoo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$as.zoo(x, y = NULL, k = 3, aggregation = "year")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TopicanalysisBundle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
