% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicanalysisBundle_class.R
\docType{data}
\name{TopicanalysisBundle}
\alias{TopicanalysisBundle}
\title{TopicanalysisBundle}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TopicanalysisBundle
}
\description{
TopicanalysisBundle
}
\section{Fields}{

\describe{
\item{\code{filenames}}{a named \code{list} of named character vectors. The names of
the list need to supply the labels/ids of the topicmodels. Names of the
vectors can be 'lda', 'labels' and ''.}

\item{\code{topicClusters}}{A named list of matrices. The list is generated by
calling the getTopicClusters-method. The matrices show the top terms of
similar topics (each column a topic). The names of the list are the labels 
of the clusters.}
}}
\section{Methods}{

\describe{
\item{\code{restore(x = NULL, verbose = TRUE)}}{Load topicmodels from disk.}
\item{\code{store(verbose = TRUE)}}{Save labels to disk (using filenames).}
\item{\code{assignLabels()}}{Assign the labels/names of the clusters in the topicClusters-field to the Topicanalysis objects.}
}
}
\examples{
\dontrun{
}
}
\keyword{datasets}

