% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TopicanalysisBundle_class.R
\docType{data}
\name{TopicanalysisBundle}
\alias{TopicanalysisBundle}
\title{TopicanalysisBundle}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x6be7010> 
 - attr(*, "name")= chr "TopicanalysisBundle_generator"
}}
\usage{
TopicanalysisBundle
}
\description{
TopicanalysisBundle
}
\examples{
\dontrun{
B <- TopicanalysisBundle$new(objects = TAs)
B$makeMatrix(big = T)
saveRDS(B$matrix, file="/Users/blaette/Lab/repos/diffusion/matrix.RData")
B$matrix <- readRDS("/Users/blaette/Lab/repos/diffusion/matrix.RData")
B$similarity <- readRDS("/Users/blaette/Lab/repos/diffusion/simil.RData")
B$doClustering()
B$getClusters(h=1)

}
}
\keyword{datasets}

