% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_method.R
\docType{class}
\name{topicCooccurrence-class}
\alias{topicCooccurrence-class}
\title{Cooccurence of Topics.}
\arguments{
\item{sizeCoi}{size of the coi}

\item{sizeRef}{size of the ref}

\item{method}{method}
}
\description{
Class required to perform the calculation of the chisquare test.
}

