% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicmodelsCmd_function.R
\name{topicmodelsCmd}
\alias{topicmodelsCmd}
\title{Generate command to run lds topicmodelling from batch.}
\usage{
topicmodelsCmd(method = "lda", infile, outfile, k)
}
\arguments{
\item{method}{either ctm or lda}

\item{infile}{input file}

\item{outfile}{output file (full path)}

\item{k}{number of topics}
}
\description{
Generate command to run lds topicmodelling from batch.
}
\examples{
\dontrun{

P <- partition("PLPRBT", speaker_type = "speech")
B <- as.speeches(
  P,
  sAttributeDates = "speaker_date", sAttributeNames = "speaker_name", gap = 500,
  mc = TRUE, progress = TRUE
)
docsToKeep <- subset(summary(B), token >= 250)[,"partition"]
B2 <- B[[docsToKeep]]

B2 <- enrich(B2, pAttribute = "word", progress = TRUE, mc = TRUE)
dtm <- as.DocumentTermMatrix(B2, pAttribute = "word")
dtmNoise <- noise(dtm, minTotal = 5, minTfIdfMean = NULL, sparse = NULL)
dtm2 <- trim(dtm, termsToDrop = unique(unlist(dtmNoise)))
emptyRows <- which(slam::row_sums(dtm2) == 0)
if (length(emptyRows) > 0) dtm2 <- dtm2[-emptyRows,]
attr(dtm2, "weighting") <- c("term frequency", "tf")
saveRDS(dtm2, file=filenames[[region]]["dtmFile"])

polmineR.topics::topicmodelsCmd(
 infile=filenames[[region]]["dtmFile"],
 outfile=file.path(projectDir, "data", paste(region, "speaker_lda_250.RData", sep="_")),
 k=nTopics
 )
}
}

